requirements.txt
fastapi
uvicorn

pydantic==2.13.2
python-dotenv==1.2.2
requests==2.33.1

openai==2.32.0

chromadb
redis

langgraph==1.1.8
langchain==1.2.15
langchain-core==1.3.0
langchain-openai==1.1.14
langchain-community
langchain-chroma
langchain-openai
langchain-anthropic   
langchain-groq

langchain-huggingface
sentence-transformers

tiktoken==0.12.0

pypdf